🔨 Move dashboard component to /dashboard
