+++ app/dashboard/CandidateProgress/index.tsx
