+++ app/dashboard/Setup/index.tsx
